Update title.json
Semantic error corrected.

The word "लेकिन" means "but" when translated but does not create a curiosity that the original video title in English is trying to incite. Using "but" is okay when people are already aware of something that your content relates to (e.g. "But what even is a V8 Engine?", here you're assuming people know about Engines and their types).
It is not useful in creating a question that people don't know the context of (and don't care about yet).